Removed utils/index and moved to explicit imports because direct importing/exporting was creating circular reference issues
