+++ src/containers/GroupForm/GroupForm.js
@@ -5,7 +5,9 @@ import i18next from 'i18next';
 import { Field, reduxForm } from 'redux-form';
 import Heading from 'd2-ui/lib/headings/Heading.component';
 import RaisedButton from 'material-ui/RaisedButton';
-import { navigateTo, asyncValidateUniqueness, asArray } from '../../utils';
+import navigateTo from '../../utils/navigateTo';
+import asyncValidateUniqueness from '../../utils/asyncValidateUniqueness';
+import asArray from '../../utils/asArray';
 import { renderSearchableGroupEditor } from '../../utils/fieldRenderers';
 import { clearItem, showSnackbar, getList } from '../../actions';
 import { NAME, USERS, MANAGED_GROUPS, FIELDS } from './config';
