+++ src/containers/RoleForm/RoleForm.js
@@ -5,7 +5,8 @@ import i18next from 'i18next';
 import { Field, reduxForm } from 'redux-form';
 import Heading from 'd2-ui/lib/headings/Heading.component';
 import RaisedButton from 'material-ui/RaisedButton';
-import { navigateTo, asyncValidateUniqueness } from '../../utils';
+import navigateTo from '../../utils/navigateTo';
+import asyncValidateUniqueness from '../../utils/asyncValidateUniqueness';
 import { clearItem, showSnackbar, getList } from '../../actions';
 import { NAME, DESCRIPTION, AUTHORITIES, FIELDS } from './config';
 import { USER_ROLE } from '../../constants/entityTypes';
