+++ src/containers/UserForm/UserForm.js
@@ -9,7 +9,9 @@ import CircularProgress from 'material-ui/CircularProgress';
 import HardwareKeyboardArrowUp from 'material-ui/svg-icons/hardware/keyboard-arrow-up';
 import HardwareKeyboardArrowDown from 'material-ui/svg-icons/hardware/keyboard-arrow-down';
 import i18next from 'i18next';
-import { navigateTo, asArray, getNestedProp } from '../../utils';
+import navigateTo from '../../utils/navigateTo';
+import asArray from '../../utils/asArray';
+import getNestedProp from '../../utils/getNestedProp';
 import api from '../../api';
 import { userFormInitialValuesSelector } from '../../selectors';
 import { clearItem, getList, showSnackbar } from '../../actions';
