+++ src/containers/UserList/UserContextMenuActions.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import i18next from 'i18next';
 import Action from 'd2-ui/lib/action/Action';
-import { navigateTo } from '../../utils';
+import navigateTo from '../../utils/navigateTo';
 import store from '../../store';
 import api from '../../api';
 import { deleteModel } from '../../utils/sharedActions';
